Move Spark options to backend-specific package
