+++ kubeflow/spark/api/spark_client_test.py
@@ -20,8 +20,8 @@ import pytest
 
 from kubeflow.common.types import KubernetesBackendConfig
 from kubeflow.spark.api.spark_client import SparkClient
+from kubeflow.spark.options import Labels
 from kubeflow.spark.test.common import FAILED, SUCCESS, TestCase
-from kubeflow.spark.types.options import Labels
 from kubeflow.spark.types.types import (
     FileJob,
     FuncJob,
