+++ kubeflow/spark/options/kubernetes.py
@@ -1,4 +1,4 @@
-# Copyright 2025 The Kubeflow Authors.
+# Copyright The Kubeflow Authors.
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
